Configuration interface for Anyfi.net on consumer Wi-Fi routers.
The recommended integration for consumer Wi-Fi routers is very similar to that
of a residential gateway, except that we explicitly refer to "Anyfi.net". The
end-user can change the 'anyfi_disabled' property on both radio frontend and
tunnel termination backend, with a single checkbox. The 'anyfi_server' property
is left under the control of the vendor, or more advanced users/administrators
with shell access. See http://anyfi.net/documentation#i_routeri for more
information.