Configuration interface for Anyfi.net on residential gateways.
To make configuration as simple as possible for the end-subscriber there is only
a single checkbox controlling the 'anyfi_disabled' property on both the
wifi-device and on the wifi-iface. This allows the service provider to set the
controller, e.g. using TR-069 or similar, and the end-subscriber to "opt-out".

Also note that the "Anyfi.net" brand is not visible to the end-subscriber;
Internet service providers typically want to deploy Anyfi.net technology under
their own brands.

See http://anyfi.net/documentation#s_rg for more information.